Converted the timing calls in nav and pid to use the new timing library. Also, dt is now in seconds for nav, pid and nav.conf instead of in microseconds. This still needs to be tested with the Stingray though.
git-svn-id: file:///home/thomas/tmp/stingray/trunk@306 ae4af862-681a-43e1-b28c-8c040b23e139

+++ libs/timing/include/timing.h
@@ -48,8 +48,8 @@ typedef struct _TIMING {
 //! Checks to see if time has elapsed.
 //! \param timer A timer value to check.
 //! \param period The amount of time to check against, in seconds.
-//! \return 1 on success, 0 on failure.
-int timing_check_period(TIMING *timer, float period);
+//! \return dt in seconds if TRUE, 0 if FALSE.
+float timing_check_period(TIMING *timer, float period);
 
 //! Sets a timer to the current system time.
 //! \param timer A timer to set.
