+++ src/common/include/pid.h
@@ -117,14 +117,14 @@ int pid_init( PID *pid, CONF_VARS *cf );
 //! \param pid PID struct.
 //! \param cf Pointer to configuration variables.
 //! \param msg Pointer to message data.
-//! \param dt Time difference from last loop run.
+//! \param dt Time difference from last loop in seconds.
 //! \param mode Which PID loop to run.
 //! \param motor_init Boolean for whether motor controller is initialized.
 void pid_loop( int pololu_fd,
                PID *pid,
                CONF_VARS *cf,
                MSG_DATA *msg,
-               int dt,
+               float dt,
                int mode,
 			   int motor_init
              );
